+++ packages/blockchain-wallet-v4-frontend/src/components/TransactionListItem/Addresses/Eth/index.js
@@ -1,0 +1,52 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import { FormattedMessage } from 'react-intl'
+
+import { Text, TextGroup } from 'blockchain-info-components'
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+`
+
+const Label = styled(Text)`
+  font-size: 13px;
+  font-weight: 300;
+`
+
+const EthAddresses = props => {
+  const { from, to } = props
+
+  return (
+    <Wrapper>
+      <TextGroup inline>
+        <Text size='13px' weight={500}>
+          <FormattedMessage
+            id='components.transactionlistitem.addresses.to'
+            defaultMessage='To: '
+          />
+        </Text>
+        <Label>{to}</Label>
+      </TextGroup>
+      <TextGroup inline>
+        <Text size='13px' weight={500}>
+          <FormattedMessage
+            id='components.transactionlistitem.addresses.from'
+            defaultMessage='From: '
+          />
+        </Text>
+        <Label>{from}</Label>
+      </TextGroup>
+    </Wrapper>
+  )
+}
+
+EthAddresses.propTypes = {
+  to: PropTypes.string.isRequired,
+  from: PropTypes.string.isRequired
+}
+
+export default EthAddresses
